pyproject.toml
[tool.poetry]
name = "report"
version = "0.1.0"
description = ""
authors = ["chaithanya2934 <chaithanya2934@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
django-celery-results = "^2.5.1"
python-dotenv = "^1.0.0"
django-getenv = "^1.3.2"
eventlet = "^0.33.3"
flower = "^2.0.0"
celery = {extras = ["redis"], version = "^5.3.1"}
polars = "^0.18.12"
connectorx = "^0.3.1"
pyarrow = "^12.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
